feat(sdk-coin-icp): enhance transaction and transfer builders; implement validation and data building for ICP transactions
TICKET: WIN-4641
